restructure , moved common libs to /lib and /include
