+++ lib/solar_monitor_server/src/solar_monitor_server.cpp
