+++ platformio.ini
@@ -12,4 +12,8 @@
 platform = espressif8266
 board = nodemcuv2
 framework = arduino
+build_flags = 
+    -I include
+
+
